+++ src/controlP5/ColorPicker.java
@@ -65,7 +65,7 @@ public class ColorPicker extends ControlGroup< ColorPicker > {
 		theControlP5.register( theControlP5.papplet , theName , this );
 	}
 
-	protected ColorPicker( ControlP5 theControlP5 , ControllerGroup< ? > theParent , String theName , int theX , int theY , int theWidth , int theHeight ) {
+	protected ColorPicker( ControlP5 theControlP5 , ControllerGroup< ? > theParent , String theName , float theX , float theY , int theWidth , int theHeight ) {
 		super( theControlP5 , theParent , theName , theX , theY , theWidth , theHeight );
 		isBarVisible = false;
 		isCollapse = false;
