+++ src/controlP5/Icon.java
@@ -52,7 +52,7 @@ public class Icon extends Controller< Icon > {
 		theControlP5.register( theControlP5.papplet , theName , this );
 	}
 
-	protected Icon( ControlP5 theControlP5 , ControllerGroup< ? > theParent , String theName , float theDefaultValue , int theX , int theY , int theWidth , int theHeight ) {
+	protected Icon( ControlP5 theControlP5 , ControllerGroup< ? > theParent , String theName , float theDefaultValue , float theX , float theY , int theWidth , int theHeight ) {
 		super( theControlP5 , theParent , theName , theX , theY , theWidth , theHeight );
 		_myValue = theDefaultValue;
 		_myCaptionLabel.align( CENTER , CENTER );
