+++ src/controlP5/Radio.java
@@ -32,7 +32,7 @@ package controlP5;
 
 public class Radio extends RadioButton {
 
-	public Radio( ControlP5 theControlP5 , ControllerGroup< ? > theParent , String theName , int theX , int theY ) {
+	public Radio( ControlP5 theControlP5 , ControllerGroup< ? > theParent , String theName , float theX , float theY ) {
 		super( theControlP5 , theParent , theName , theX , theY );
 	}
